Add script for parsing JSON Slack logs into IRC-style text logs
This utility takes a json filename as a command line argument, then parses and
reformats the file to appear like an IRC-style text log and prints it to
standard output.

Note that this script requires users.json and channels.json files which
describe user/channel ID mappings to be located in the current directory in
order to run.